Eliminate compile warning on Windows
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@85010 13f79535-47bb-0310-9956-ffa450edef68

+++ include/httpd.h
@@ -964,7 +964,7 @@ int strcasecmp(const char *a, const char *b);
 #endif
 
 #ifndef HAVE_STRNCASECMP
-int strncasecmp(const char *a, const char *b, int n);
+int strncasecmp(const char *a, const char *b, size_t n);
 #endif
 
 /* Regexes */
